+++ src/create/index.js
@@ -36,7 +36,7 @@ export default () => {
       `plugin:${packageName`eslint-plugin-vue`}/vue3-recommended`,
       `plugin:${packageName`eslint-plugin-prettier`}/recommended`,
     ],
-    globals: { self: true, window: true },
+    globals: { globalThis: true, self: true, window: true },
     overrides: [
       {
         files: '**/*.spec.js',
@@ -91,10 +91,14 @@ export default () => {
       'no-console': 'off',
       'no-constant-condition': ['error', { checkLoops: false }],
       'no-continue': 'off',
+      'no-lonely-if': 'off',
       'no-negated-condition': 'error',
+      'no-nested-ternary': 'off',
       'no-param-reassign': 'off',
       'no-promise-executor-return': 'off',
+
       'no-regex-spaces': 'off',
+
       // https://github.com/facebook/create-react-app/issues/12847
       'no-restricted-globals': [
         'error',
@@ -113,7 +117,6 @@ export default () => {
           name: g,
         })),
       ],
-
       'no-restricted-imports': ['error', { paths: eslintRestrictedImports }],
       'no-restricted-syntax': ['error', "LogicalExpression[operator='??']"],
       'no-return-assign': 'off',
@@ -137,7 +140,6 @@ export default () => {
         { blankLine: 'always', next: 'export', prev: '*' },
       ],
       'prefer-arrow/prefer-arrow-functions': ['error'],
-      'prefer-destructuring': 'off',
       [`${packageName`prettier`}/prettier`]: [
         'error',
         {
@@ -147,6 +149,7 @@ export default () => {
           trailingComma: 'all',
         },
       ],
+      'prefer-destructuring': 'off',
       'promise/prefer-await-to-then': 'error',
       'require-await': 'error',
       'simple-import-sort/imports': 'error',