+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [5.0.13](https://github.com/dword-design/eslint-config/compare/v5.0.12...v5.0.13) (2024-12-15)
+
+
+### Bug Fixes
+
+* add globalThis to globals ([bfccaaa](https://github.com/dword-design/eslint-config/commit/bfccaaa58a872dc7901c64b16968ee98e42a6808))
+* fix this ([7587163](https://github.com/dword-design/eslint-config/commit/7587163a14eef6fedd763dcff45150dd0ad6b03d))
+
+## [5.0.12](https://github.com/dword-design/eslint-config/compare/v5.0.11...v5.0.12) (2024-11-17)
+
+
+### Bug Fixes
+
+* disable lonely ifs rule ([c5dc68a](https://github.com/dword-design/eslint-config/commit/c5dc68a41a4c75a2a9fe0b301cdef72caa8ca6b2))
+
 ## [5.0.11](https://github.com/dword-design/eslint-config/compare/v5.0.10...v5.0.11) (2024-08-13)